requirements.txt
PyQt5==5.15.7
RPi.GPIO==0.7.1
